% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{create_configuration}
\alias{create_configuration}
\title{Creates a new configuration for the specified configuration name}
\usage{
create_configuration(EngineType = NULL, EngineVersion = NULL,
  Name = NULL)
}
\arguments{
\item{EngineType}{Required. The type of broker engine. Note: Currently, Amazon MQ supports only ACTIVEMQ.}

\item{EngineVersion}{Required. The version of the broker engine. Note: Currently, Amazon MQ supports only 5.15.6 and 5.15.0.}

\item{Name}{Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.}
}
\description{
Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).
}
\section{Accepted Parameters}{
\preformatted{create_configuration(
  EngineType = "ACTIVEMQ",
  EngineVersion = "string",
  Name = "string"
)
}
}

