% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{create_tags}
\alias{create_tags}
\title{Add a tag to a resource}
\usage{
create_tags(ResourceArn, Tags = NULL)
}
\arguments{
\item{ResourceArn}{[required] the Amazon Resource Name (ARN)}

\item{Tags}{The key-value pair for the resource tag.}
}
\description{
Add a tag to a resource.
}
\section{Accepted Parameters}{
\preformatted{create_tags(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

