% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Remove a tag from a resource}
\usage{
delete_tags(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] the Amazon Resource Name (ARN)}

\item{TagKeys}{[required] An array of tag keys to delete}
}
\description{
Remove a tag from a resource.
}
\section{Accepted Parameters}{
\preformatted{delete_tags(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

