% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{describe_user}
\alias{describe_user}
\title{Returns information about an ActiveMQ user}
\usage{
describe_user(Username, BrokerId)
}
\arguments{
\item{Username}{The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.}

\item{BrokerId}{The unique ID that Amazon MQ generates for the broker.}
}
\description{
Returns information about an ActiveMQ user.
}
\examples{

}
