% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{list_configuration_revisions}
\alias{list_configuration_revisions}
\title{Returns a list of all revisions for the specified configuration}
\usage{
list_configuration_revisions(ConfigurationId, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ConfigurationId}{The unique ID that Amazon MQ generates for the configuration.}

\item{MaxResults}{The maximum number of configurations that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.}

\item{NextToken}{The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.}
}
\description{
Returns a list of all revisions for the specified configuration.
}
\examples{

}
