% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{list_tags}
\alias{list_tags}
\title{Lists tags for a resource}
\usage{
list_tags(
  ResourceArn = "string"
)
}
\arguments{
\item{ResourceArn}{[required] the Amazon Resource Name (ARN)}
}
\description{
Lists tags for a resource.
}
\examples{

}
