% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_package.R
\docType{package}
\name{paws.mq-package}
\alias{paws.mq}
\alias{paws.mq-package}
\title{paws.mq: AmazonMQ}
\description{
Amazon MQ is a managed message broker service for
Apache ActiveMQ that makes it easy to set up and operate message
brokers in the cloud. A message broker allows software applications
and components to communicate using various programming languages,
operating systems, and formal messaging protocols.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
