% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{reboot_broker}
\alias{reboot_broker}
\title{Reboots a broker}
\usage{
reboot_broker(BrokerId)
}
\arguments{
\item{BrokerId}{[required] The unique ID that Amazon MQ generates for the broker.}
}
\description{
Reboots a broker. Note: This API is asynchronous.
}
\section{Accepted Parameters}{
\preformatted{reboot_broker(
  BrokerId = "string"
)
}
}

