% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{update_broker}
\alias{update_broker}
\title{Adds a pending configuration change to a broker}
\usage{
update_broker(BrokerId, AutoMinorVersionUpgrade = NULL,
  Configuration = NULL, EngineVersion = NULL, Logs = NULL)
}
\arguments{
\item{BrokerId}{The name of the broker. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain whitespaces, brackets, wildcard characters, or special characters.}

\item{AutoMinorVersionUpgrade}{Enables automatic upgrades to new minor versions for brokers, as Apache releases the versions. The automatic upgrades occur during the maintenance window of the broker or after a manual broker reboot.}

\item{Configuration}{A list of information about the configuration.}

\item{EngineVersion}{The version of the broker engine. Note: Currently, Amazon MQ supports only 5.15.6 and 5.15.0.}

\item{Logs}{Enables Amazon CloudWatch logging for brokers.}
}
\description{
Adds a pending configuration change to a broker.
}
\examples{

}
