% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{update_configuration}
\alias{update_configuration}
\title{Updates the specified configuration}
\usage{
update_configuration(ConfigurationId, Data = NULL, Description = NULL)
}
\arguments{
\item{ConfigurationId}{[required] The unique ID that Amazon MQ generates for the configuration.}

\item{Data}{Required. The base64-encoded XML configuration.}

\item{Description}{The description of the configuration.}
}
\description{
Updates the specified configuration.
}
\section{Accepted Parameters}{
\preformatted{update_configuration(
  ConfigurationId = "string",
  Data = "string",
  Description = "string"
)
}
}

