% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{update_user}
\alias{update_user}
\title{Updates the information for an ActiveMQ user}
\usage{
update_user(Username, BrokerId, ConsoleAccess = NULL, Groups = NULL,
  Password = NULL)
}
\arguments{
\item{Username}{Required. The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.}

\item{BrokerId}{The unique ID that Amazon MQ generates for the broker.}

\item{ConsoleAccess}{Enables access to the the ActiveMQ Web Console for the ActiveMQ user.}

\item{Groups}{The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.}

\item{Password}{The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas.}
}
\description{
Updates the information for an ActiveMQ user.
}
\examples{

}
