% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{describe_configuration}
\alias{describe_configuration}
\title{Returns information about the specified configuration}
\usage{
describe_configuration(ConfigurationId)
}
\arguments{
\item{ConfigurationId}{The unique ID that Amazon MQ generates for the configuration.}
}
\description{
Returns information about the specified configuration.
}
\examples{

}
