% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mq_operations.R
\name{describe_configuration_revision}
\alias{describe_configuration_revision}
\title{Returns the specified configuration revision for the specified configuration}
\usage{
describe_configuration_revision(ConfigurationId, ConfigurationRevision)
}
\arguments{
\item{ConfigurationId}{[required] The unique ID that Amazon MQ generates for the configuration.}

\item{ConfigurationRevision}{[required] The revision of the configuration.}
}
\description{
Returns the specified configuration revision for the specified configuration.
}
\section{Accepted Parameters}{
\preformatted{describe_configuration_revision(
  ConfigurationId = "string",
  ConfigurationRevision = "string"
)
}
}

